refactor: move fetured section in a separate component
