+++ src/components/magicui/bento-grid.tsx
@@ -18,7 +18,7 @@ interface BentoCardProps extends ComponentPropsWithoutRef<"div"> {
   Icon: React.ElementType;
   description: string;
   href: string;
-  cta: string;
+  cta?: string;
   modal?: boolean;
 }
 
@@ -73,42 +73,43 @@ const BentoCard = ({
         "pointer-events-none absolute bottom-0 flex w-full translate-y-10 transform-gpu flex-row items-center p-4 opacity-0 transition-all duration-300 group-hover:translate-y-0 group-hover:opacity-100"
       )}
     >
-      {modal ? (
-        <Dialog>
+      {cta &&
+        (modal ? (
+          <Dialog>
+            <Button
+              variant="ghost"
+              asChild
+              size="sm"
+              className="pointer-events-auto"
+            >
+              <DialogTrigger>
+                {cta}
+                <ArrowRightIcon className="ms-2 h-4 w-4 rtl:rotate-180" />
+              </DialogTrigger>
+            </Button>
+            <DialogContent>
+              <div className="bg-black p-4 rounded-xl shadow-lg">
+                <iframe
+                  src="https://www.linkedin.com/embed/feed/update/urn:li:activity:7303170251662200833/"
+                  className="h-[800px] w-full rounded-lg"
+                  title="Embedded post"
+                ></iframe>
+              </div>
+            </DialogContent>
+          </Dialog>
+        ) : (
           <Button
             variant="ghost"
             asChild
             size="sm"
             className="pointer-events-auto"
           >
-            <DialogTrigger>
-              See more
+            <a href={href} target="_blank">
+              {cta}
               <ArrowRightIcon className="ms-2 h-4 w-4 rtl:rotate-180" />
-            </DialogTrigger>
+            </a>
           </Button>
-          <DialogContent>
-            <div className="bg-black p-4 rounded-xl shadow-lg">
-              <iframe
-                src="https://www.linkedin.com/embed/feed/update/urn:li:activity:7303170251662200833/"
-                className="h-[800px] w-full rounded-lg"
-                title="Embedded post"
-              ></iframe>
-            </div>
-          </DialogContent>
-        </Dialog>
-      ) : (
-        <Button
-          variant="ghost"
-          asChild
-          size="sm"
-          className="pointer-events-auto"
-        >
-          <a href={href} target="_blank">
-            {cta}
-            <ArrowRightIcon className="ms-2 h-4 w-4 rtl:rotate-180" />
-          </a>
-        </Button>
-      )}
+        ))}
     </div>
     <div className="pointer-events-none absolute inset-0 transform-gpu transition-all duration-300 group-hover:bg-black/[.03] group-hover:dark:bg-neutral-800/10" />
